pull joiner out
