+++ pyxmex/__init__.py
@@ -1,1 +1,2 @@
 from pyxmex.parser import Parser
+from pyxmex.joiner import Joiner
